videosaur/requirements_transition_py312.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array_record==0.8.3
arrow==1.4.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.14.2
black==25.11.0
blake3==1.0.8
bleach==6.3.0
braceexpand==0.1.7
brotlicffi
cachetools==6.2.1
cbor2==5.7.1
certifi
cffi
cfgv==3.5.0
charset-normalizer
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
compressed-tensors==0.11.0
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
depyf==0.19.0
diffusers==0.35.2
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dm-tree==0.1.9
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
entrypoints==0.4
etils==1.13.0
eval_type_backport==0.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.120.4
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
fastrlock==0.8.3
filelock
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec
gast==0.6.0
gdown==5.2.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
gmpy2
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutabledict==4.2.2
importlib_metadata
importlib_resources==6.5.2
iniconfig==2.3.0
interegular==0.3.3
ipykernel==7.0.1
ipython==9.6.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2
jiter==0.11.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==3.12.0
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.15.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
mistune==3.1.4
mkl-service==2.5.2
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess
mypy_extensions==1.1.0
namex==0.1.0
nbclassic==1.3.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx
ninja==1.13.0
nodeenv==1.9.1
notebook==7.5.0
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.10.2.21
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
openai==2.6.1
openai-harmony==0.0.4
opencv-python-headless==4.12.0.88
opentelemetry-api
opt_einsum==3.4.0
optree==0.18.0
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pathos
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow
pkgutil_resolve_name==1.3.10
platformdirs==4.5.0
pluggy==1.6.0
pox==0.3.6
ppft==1.7.7
pre_commit==4.5.0
proglog==0.1.12
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycountry==24.6.1
pycparser
pydantic==2.12.2
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing
pyrsistent==0.20.0
PySocks
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytokens==0.3.0
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML
pyzmq==27.1.0
ray==2.51.1
referencing==0.37.0
regex==2025.10.23
requests
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.3
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.7
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.42.0
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
starlette==0.49.3
submitit
sympy
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.4
tensorflow-datasets==4.9.9
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
tensorflow_cpu==2.20.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
timm==1.0.21
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.8.0
torch-geometric==2.6.1
torchaudio==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.4.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions
tzdata==2025.2
uri-template==1.3.0
urllib3
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
vllm==0.11.0
wandb==0.22.2
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
zipp
